Refactor UseCase and Root UoW traits with macros
Consolidated operations in UseCase and Root Unit of Work traits by introducing macros to automatically generate actions like create, update, delete, and relationship management. Removed redundant code and moved towards a declarative approach for cleaner and more maintainable code. Added support for handling relationship operations within UseCase via new get and set use case relationship modules.

+++ crates/direct_access/src/use_case/use_cases/remove_use_case_multi_uc.rs
@@ -1,4 +1,4 @@
-use super::common::UseCaseUnitOfWorkFactoryTrait;
+use super::UseCaseUnitOfWorkFactoryTrait;
 use anyhow::{Ok, Result};
 use common::types::Savepoint;
 use common::{types::EntityId, undo_redo::UndoRedoCommand};
@@ -23,6 +23,17 @@ impl RemoveUseCaseMultiUseCase {
         let mut uow = self.uow_factory.create();
         uow.begin_transaction()?;
         let savepoint = uow.create_savepoint()?;
+        // check if id exists
+        let mut exists = true;
+        for id in ids {
+            if uow.get_use_case(id)?.is_none() {
+                exists = false;
+                break;
+            }
+        }
+        if !exists {
+            return Err(anyhow::anyhow!("One or more ids do not exist"));
+        }
         uow.delete_use_case_multi(ids)?;
         uow.commit()?;
